+++ fixtures/fragments/slugify.md
@@ -1,0 +1,27 @@
+# #️⃣ b
+
+# ☔️ c
+
+# 🅰️ d
+
+# foo 1
+
+# foo
+
+# foo
+
+# foo
+
+# foo 1
+
+# a *b* c
+
+# a `b` c
+
+# a `#` c
+
+# À, Á, Â, Ã, Ä, Å or à, á, â, ã, ä, å
+
+# an &mdash; emdash
+
+# 🏛️ MIT 许可证
