+++ lychee-lib/src/extract/mod.rs
@@ -77,7 +77,7 @@ mod tests {
     use pretty_assertions::assert_eq;
     use reqwest::Url;
     use std::{collections::HashSet, path::Path};
-    use test_utils::{fixtures_path, load_fixture, mail, website};
+    use test_utils::{load_fixture, mail, website};
 
     use super::*;
     use crate::{
