+++ test-utils/src/lib.rs
@@ -133,7 +133,7 @@ macro_rules! fixtures_path {
 #[macro_export]
 macro_rules! load_fixture {
     ($filename:expr) => {{
-        let path = fixtures_path!().join($filename);
+        let path = test_utils::fixtures_path!().join($filename);
         std::fs::read_to_string(path).unwrap()
     }};
 }
@@ -148,7 +148,7 @@ macro_rules! fixture_uri {
     ($subpath:expr) => {{
         use url::Url;
         let fixture_url =
-            Url::from_directory_path(fixtures_path!()).expect("fixture path should be a valid URL");
+            Url::from_directory_path(test_utils::fixtures_path!()).expect("fixture path should be a valid URL");
 
         // joining subpath onto a Url allows the subpath to contain a fragment
         fixture_url
